added 1Up powerup

+++ src/vvvvvv/vvvvvv.py
@@ -1,5 +1,6 @@
 from Obstacle import Obstacle
 from Player import Player
+from OneUP import OneUP
 import pygame
 import random
 
@@ -25,6 +26,7 @@ white = (255, 255, 255)
 grey = (150, 150, 150)
 
 obstacles = []
+powerups = []
 
 def write_score():
     font = pygame.font.Font(None, 24)
@@ -39,23 +41,29 @@ def generate_obstacle():
     if p < 1:
         obstacles.append(Obstacle())
 
-def check_collision(obstacle):
-    obstacle_rect = obstacle.image.get_rect()
-    (x, y) = obstacle.pos
-    obstacle_rect.top = y
-    obstacle_rect.left = x
+def generate_oneup():
+    p = random.randint(0, 1000)
+    if p < 1:
+        powerups.append(OneUP())
+
+def check_collision(other):
+    other_rect = other.image.get_rect()
+    (x, y) = other.pos
+    other_rect.top = y
+    other_rect.left = x
 
     player_rect = player.image.get_rect()
     (x, y) = player.pos
     player_rect.top = y
     player_rect.left = x
 
-    return obstacle_rect.colliderect(player_rect)
+    return other_rect.colliderect(player_rect)
 
 while not player.dead:
     screen.fill((0, 0, 0))
     write_score()
     generate_obstacle()
+    generate_oneup()
 
     # check collision with bounds
     (_, y) = player.pos
@@ -76,8 +84,23 @@ while not player.dead:
     # check collision with obstacles
     for obstacle in obstacles:
         if check_collision(obstacle):
-            player.kill()
-            break
+            obstacles.remove(obstacle)
+            player.onedown()
+            if player.extra_lifes < 0:
+                player.kill()
+                break
+
+    # check collision with powerups
+    for powerup in powerups:
+        if check_collision(powerup):
+            player.oneup()
+            powerups.remove(powerup)
+
+    # check if any powerups must be removed
+    for powerup in powerups:
+        remove = powerup.decrease_ticks()
+        if remove:
+            powerups.remove(powerup)
 
     # check pressed buttons
     for event in pygame.event.get():
@@ -107,6 +130,9 @@ while not player.dead:
             obstacles.remove(obstacle)
         else:
             screen.blit(obstacle.image, obstacle.pos)
+
+    for powerup in powerups:
+        screen.blit(powerup.image, powerup.pos)
 
     pygame.display.flip()
     clock.tick(60)
